Controllerの機能追加
*削除エラーチェック追加

+++ backend/src/controllers/clientController.js
@@ -27,6 +27,51 @@ class ClientController {
       res.status(500).send();
     }
   }
+  /**
+   * 取引先情報詳細取得
+   *
+   * @param {*} req リクエスト情報
+   * @param {*} res レスポンス情報
+   */
+  async findByCode(req, res) {
+    // パスパラメータチェック
+    // パスパラメータの取引先コードが設定されていない場合
+    const errors = [];
+    if (!req.params.clientCode) {
+      errors.push({ field: "clientCode", message: "取引先コードが設定されていません" });
+    }
+
+    // パスパラメータの取引先コードが8桁で設定されていない場合
+    else if (req.params.clientCode.length !== 8) {
+      errors.push({ field: "clientCode", message: "取引先コードは8桁で設定してください" });
+    }
+
+    // パスパラメータの取引先コードが半角英数で設定されていない場合
+    else if (!/^[A-Za-z0-9]+$/.test(req.params.clientCode)) {
+      errors.push({ field: "clientCode", message: "取引先コードは半角英数で設定してください" });
+    }
+
+    if (errors.length > 0) {
+      console.log(errors);
+      return res.status(400).json({ errors: errors });
+    }
+
+    try {
+      const client = await clientService.findByCode(req.params.clientCode);
+      res.json(client);
+    } catch (err) {
+      console.error(err);
+
+      if (err instanceof NotFoundError) {
+        // 存在チェックエラー
+        errors.push({ field: "clientCode", message: "この取引先情報は存在しません" });
+        console.log(errors);
+        res.status(NotFoundError.status).json({ errors: errors });
+      } else {
+        res.status(500).send();
+      }
+    }
+  }
 
   /**
    * 取引先情報削除
@@ -35,32 +80,40 @@ class ClientController {
    * @param {*} res レスポンス情報
    */
   async delete(req, res) {
+     // パスパラメータチェック
+    // パスパラメータの取引先コードが設定されていない場合
+    const errors = [];
+    if (!req.params.clientCode) {
+      errors.push({ field: "clientCode", message: "取引先コードが設定されていません" });
+    }
+
+    // パスパラメータの取引先コードが8桁で設定されていない場合
+    else if (req.params.clientCode.length !== 8) {
+      errors.push({ field: "clientCode", message: "取引先コードは8桁で設定してください" });
+    }
+
+    // パスパラメータの取引先コードが半角英数で設定されていない場合
+    else if (!/^[A-Za-z0-9]+$/.test(req.params.clientCode)) {
+      errors.push({ field: "clientCode", message: "取引先コードは半角英数で設定してください" });
+    }
+
+    if (errors.length > 0) {
+      console.log(errors);
+      return res.status(400).json({ errors: errors });
+    }
+
+
     try {
-      // パスパラメータチェック
-      // パスパラメータの取引先コードが設定されていない場合
-      if (!req.params.clientCode) {
-        console.error(err);
-        res.status(400).send();
-      }
-
-      // パスパラメータの取引先コードが8桁で設定されていない場合
-      else if (req.params.clientCode.length !== 8) {
-        console.error(e);
-        res.status(400).send();
-      }
-
-      // パスパラメータの取引先コードが半角英数で設定されていない場合
-      else if (!/^[A-Za-z0-9]+$/.test(req.params.clientCode)) {
-        console.error(err);
-        res.status(400).send();
-      }
       await clientService.delete(req.params.clientCode);
       res.send();
-    } catch (e) {
-      console.log(e);
+    } catch (err) {
+      console.log(err);
 
-      if (e instanceof NotFoundError) {
-        res.status(NotFoundError.status).send();
+      if (err instanceof NotFoundError) {
+        // 存在チェックエラー
+        errors.push({ field: "clientCode", message: "この取引先情報は存在しません" });
+        console.log(errors);
+        res.status(NotFoundError.status).json({errors:errors});
       } else {
         res.status(500).send();
       }
